Switch icons to 'highlighter' symbol
This adds the new 'highlighter' symbol to the five CodeMirror SVG
icons, and regenerates their PNG counterparts (using Inkscape).

The only changes to the SVG files apart from the paths (and their
positions) are whitespace formatting. I'm not sure all those
gradients need to be there, but I guess someone did that for a
reason. :-)

Bug: T164441
Change-Id: Ibdb8ecf53eb03fb1d1805e788a3e497e4941263a