Reload CodeMirror to fix a caret issue
A common issue, reported in VPI as Possible Thursday weirdness, with
source editor users using syntax highlighting is that the caret can
often be misplaced as to where the user selects. This PR adds code
to reload CM when loaded to adjust for this behavior.
Moved to after CM has finished initialising. Tested working.
Add another instance of codeMirror.refresh().

Bug: T305333
Change-Id: I9a81ffd41a3ba1321b7b5744ba096583cbb1d96c