CodeMirrorModeMediaWiki: tokenizer reworked
A new implementation for the bold and italic apostrophes, which fixes existing bugs associated with unbalanced apostrophes (such as highlighting ` ''' ''` or `''[[a]]'''`) and improves performance by tokenizing only once. This new implementation should also be more reliable considering customized tag modes in the future. It has been extensively tested using the parser tests from the PHP parser.

This patch also permanently fixes `Unknown highlighting tag ` warnings which may appear unexpectedly.

This patch includes a new test case related to T366035. The goal is to make sure that the parser will not crash with a very long line.

Bug: T304567
Change-Id: I966d452a7dfd8901d2bb92b2a8d191d024da5ac9