CodeMirrorKeymap: system for adding key bindings and showing help dialog
Inspired by the 2017 editor, add a help dialog accessible with
Ctrl-Shift-/ that documents noteworthy key bindings. This uses a Codex
Dialog component. Also document modifier bindings, such as Mod + click
to open links. The help dialog is also accessible from a new
'keyboard shortcuts' link next to the 'help' link the preferences panel.

Mod-/ is already reserved for commenting, which now uses
jQuery.textSelection instead of CM's internal CommentTokens interface.
The native CM historyKeymap is re-implemented by CodeMirrorKeymap.

Allow integrations such as CodeMirrorMediaWiki to dynamically add and
document their own key bindings, using an interface that is roughly
compatible with CM's internal KeyBinding. This uses the same
rudimentary browser-detection logic as the CodeMirror lib, dropping the
need for the jQuery.client dependency.

Key bindings that are registered with the same name as the entry in
CodeMirrorPreferences' extensionRegistry are shown or hidden from the
help dialog according to the user's preference.

CodeMirrorMediaWikiKeymap is introduced for MW-specific key bindings and
wikitext formatting. The implementations live here in CodeMirror as it
is meant to work without WikiEditor (or any other editor). Some
implementations differ slightly from WikiEditor, such as the heading
shortcuts (i.e. Mod+3). When used on an existing heading, it will
replace only the heading syntax.

The CM implementation of jQuery.textSelection 'encapsulateSelection' has
been changed so that the 'pre' and 'post' text are not applied to empty
lines when 'splitLines' is used. This is to be consistent with the 2017
editor.

Several new mw.hooks are added for both public and internal consumption
(these intentionally are not marked as @stable yet):
* ext.CodeMirror.preferences.ready – Called at the end of the
  CodeMirrorPreferences constructor, giving callers access to instance
  properties and methods.
* ext.CodeMirror.preferences.change – fired when a CM preference is
  changed. Used by CodeMirrorKeymap to update display when applicable
  extensions are disabled.
* ext.CodeMirror.preferences.display (internal) – Fired just before the
  preferences panel is displayed. This is used by CodeMirrorKeymap to
  add the click listener to the 'keyboard shortcuts' link in the
  preferences panel.

Other changes:
* Fix JSDoc eslint warnings by specifying known types that we don't
  necessarily want to require() just to appease the linting gods.

TODOs for follow-up patches:
* Move panel and dialog HTML generation to Mustache templates.

Bug: T380913
Change-Id: I26fac8048b4aa740abdecb2f081348cd2090afa3