Hooks: ensure GetPreferences hook and 2017 editor respect beta feature
For users of the 2003 editor (or if WikiEditor is not available), the
user is supposed to still be able to toggle CodeMirror on and off using
the visible preference in the new 'Syntax highlighting' section of
Special:Preferences. This patch fixes that, restoring use of standalone
CodeMirror for those who have the beta feature on but WikiEditor off.

This patch also fixes a bug that prevented 2017 editor users from
getting CM6 via the beta feature.

Bug: T376735
Change-Id: I69562ee1f936050301900ee80ab1aa8eaf92ec0d