CodeMirrorTextSelection: add support for splitlines in encapsulate
This further develops the jQuery.textSelection implementation.
The 'splitlines' option of the encapsulateSelection method, combined
with 'pre' or 'post', will wrap each line with 'pre' and 'post'.

This is needed to add a keyboard shortcut for preformatted text, which
adds a single space to beginning of each line of the selected text.

Unlike the native jQuery.textSelection, after 'splitlines' is used we
select all altered text. A TODO is left to make note of this
discrepancy and possibly upstream it to jQuery.textSelection.

Bug: T380913
Change-Id: I9076e8559e2f6f28a8a69262f8676669577fd69a